Implemented hierarchy cache - delete has problems
Fixed styling

Added commit

Reverted to old $ttl check

Fixed styling

Fixed a small part of the delete problem

Setting binary protocol in constructor now

Styling fixed

Bumped taggable cache to 0.4

 Date:      Thu Jan 14 14:56:06 2016 +0100
	modified:   composer.json
	modified:   src/MemcachedCachePool.php
	new file:   tests/CreatePoolTrait.php
        new file:   tests/IntegrationHierarchyTest.php
 	modified:   tests/IntegrationPoolTest.php
	modified:   tests/IntegrationTagTest.php
Implemented hierarchy cache
